+++ CHANGELOG.md
@@ -1060,3 +1060,19 @@ should use 4.0.1-alpha.0 for testing.
 #### web3-validator
 
 -   `build` entry from `package.json` (#5755)
+
+### Fixed
+
+#### web3-eth
+
+-   Enable transaction with local wallet index in the `to` field (#5731)
+
+#### web3-eth
+
+-   Add error for invalid transaction receiver (#5731)
+
+### Added
+
+#### web3-types
+
+-   Add `TransactionWithFromLocalWalletIndex`, `TransactionWithToLocalWalletIndex` and `TransactionWithFromAndToLocalWalletIndex` types (#5731)